test: Move QuantDatapoint into new test module
QuantDatapoint is a general class used troughout all "quant" modules.
Since it uses the intermediate (standardized) data format for creating
datapoint dataframes, it doesn't need to be tested with the output of
different software tools.
